Rename config_client.methods to config_client.interfaces
