+++ src/aiovantage/config_client/README.md
@@ -21,7 +21,6 @@ object types, available in [`objects`](objects).
 Additionally, a few helper functions for fetching objects are provided by
 [`helpers.py`](helpers.py).
 
-
 ## Examples
 
 ### Lookup objects by type, using a helper
@@ -44,23 +43,21 @@ async with ConfigClient("hostname") as client:
     objects = get_objects_by_id(client, [118])
 ```
 
-
 ### Make a request using a method class
 
 ```python
 from aiovantage.config_client import ConfigClient
-from aiovantage.config_client.methods.introspection import GetVersion
+from aiovantage.config_client.interfaces.introspection import GetVersion
 
 async with ConfigClient("hostname") as client:
     version = await client.request(GetVersion)
 ```
 
-
 ### Make a request using a method class, with params
 
 ```python
 from aiovantage.config_client import ConfigClient
-from aiovantage.config_client.methods.login import Login
+from aiovantage.config_client.interfaces.login import Login
 
 async with ConfigClient("hostname") as client:
     await client.request(Login, Login.Params("username", "password"))
