+++ src/aiovantage/config_client/__init__.py
@@ -29,9 +29,9 @@ from xsdata.formats.dataclass.parsers.handlers import XmlEventHandler
 from xsdata.formats.dataclass.serializers import XmlSerializer
 from xsdata.formats.dataclass.serializers.config import SerializerConfig
 
-from aiovantage.config_client.methods import Call, Method, Return
-from aiovantage.config_client.methods.introspection import GetVersion
-from aiovantage.config_client.methods.login import Login
+from aiovantage.config_client.interfaces import Call, Method, Return
+from aiovantage.config_client.interfaces.introspection import GetVersion
+from aiovantage.config_client.interfaces.login import Login
 from aiovantage.connection import BaseConnection
 from aiovantage.errors import ClientResponseError, LoginFailedError, LoginRequiredError
