+++ src/aiovantage/config_client/interfaces/__init__.py
