+++ src/aiovantage/config_client/interfaces/configuration/close_filter.py
