+++ src/aiovantage/config_client/interfaces/configuration/open_filter.py
