+++ src/aiovantage/config_client/interfaces/types.py
